Many fixes
Discovered an issue where the player and the enemy would only cause the damage that was set by the bullet. Now each enemy can have its own damage modifier due to the enemy and the player having separate bullet scripts. It was also discovered that every weapon the player picked up had set damage on them as well due to the bullet having a set damage amount. Now every weapon fires a separate damage amount. Adjusted the shoot speed animation of the enemy as 1 on the shoot speed fires a 3 round burst due to the animation length. Upped the player health. Decreased the shoot rate and speed of the enemy robot soldier. Updates will need to be done to the default enemy.