Gun updates
The guns now instantiate a bullet from the barrel as opposed to our previous generic location. Every guns bullets now track to the center of the reticle using a ray cast to obtain the location based on what the player sees. On top of that the guns now have a muzzle flash.