Gun Update
Shotgun can now be picked up and used. Fixed the mesh of the shotgun so that it was scaled correctly. Fixed an orientation issue with the meshes of the shotgun. Fixed an issue where the bullets fired from a gun were set to a fixed speed and not what was set per the scriptable object. Fixed an issue where the bullets would pass through a wall. Fixed an issue where the bullets would not shoot in the correct direction if the player was looking down.