Shotgun, Player, UI, Aiming logic, M4 updates
The original shotgun model had a meshed sight that could not be rendered over. I took the model into blender removed the faces of the sight then re-uploaded into Unity. Made a shotgun sight image in the UI made reference to it in game manager then attached the image. Updated the player so that it has the holder from the Gun branch due to the need to set the transform location of the shotgun. Updated the aim down sights logic to implement a time delay to allow for a more realistic aiming feature.