Trap Overhaul and Level 2 Fixes
Added Various Sfx to the mp3 file folder. Added spike and laser trap sfx. Changed the material on pressure plate and platforming sections. Added explosion particles for the mine when exploding. I have not tied the spike and laser sfx to the audio slider yet. Waiting for everyone else to finish in the audiocontrol and playercontroller scripts.