Ammo Box Bug Fix
I noticed at the top of trap house (where the key is) at the wall where the 2 ammo boxes and 1 med box is. If you continuously press F to interact with any of the 3 boxes you will endlessly get more ammo and health. It also looks like ammo resets after a certain limit (after 165). I fixed all the ammo med and ammo boxes in the level so you can not get infinite ammo